# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kasir Boy API",
        "description": "API Documentation for Kasir Boy Backend",
        "contact": {
            "email": "muhfawas7@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/admin/landing-configurations": {
            "get": {
                "tags": [
                    "Landing Configuration"
                ],
                "summary": "Get landing page configuration",
                "operationId": "e0937861e485836ba82dc3fb3344ec50",
                "responses": {
                    "200": {
                        "description": "Landing configuration retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Landing Configuration"
                ],
                "summary": "Update landing page configuration",
                "operationId": "41b3a1af5f625933a97f9b6187f90528",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "hero_image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "hero_description": {
                                        "type": "string"
                                    },
                                    "about_image_1": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "about_image_2": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "about_image_3": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "about_image_4": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "about_description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Landing configuration updated successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}